backend/requirements.txt
# Flask
Flask==3.0.3
flask-cors==4.0.1
 
# Banco de dados
mysql-connector-python==9.6.0
 
# Upload de arquivos
Werkzeug==3.0.3
 